+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/repository/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends ListCrudRepository<User, Integer> {
+public interface UserRepository extends ListCrudRepository<User, Long> {
     Optional<User> findByLogin(String login);
 
     boolean existsByLoginAndPassword(String login, String password);
